// frontend-ui/src/api/models/TicketType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Event } from './Event';
export type TicketType = {
    id?: number;
    name?: string;
    event?: Event;
    price?: number;
    quota?: number;
    purchaseLimit?: number;
    startSale?: string;
    endSale?: string;
};